Create run-tests.yml (#1)
* Create run-tests.yml

Can I run the tests for this project at no/minor cost without having to do any actual work because it's a Sunday morning and I don't have the inclination to try very hard? Not to mention the kids are in the bath, so I am under time constraints too.